requirements.txt
python-dotenv
requests
jsonpath_rw_ext
urllib3
boto3
